start containers in pod

+++ fabric/start-java8.sh
@@ -2,10 +2,29 @@
 if test "$eula" = "true" ; then
 	sed -i 's/eula=false/eula=true/' eula.txt
 fi
-# rm -r ./versions
-# cp -r ../versions ./versions
-rm -r ./libraries
-cp -r ../libraries ./libraries
+if test -n "$seed" ; then
+	sed -i 's/level-seed=.*/level-seed=$seed/' server.properties
+fi
+if test -n "$rcon" ; then
+	sed -i 's/enable-rcon=.*/enable-rcon=$rcon/' server.properties
+fi
+if test -n "$rconpwd" ; then
+	sed -i 's/rcon.password=.*/rcon.password=$rconpwd/' server.properties
+fi
+if test -n "$gamemode" ; then
+	sed -i 's/gamemode=.*/gamemode=$gamemode/' server.properties
+fi
+if test -n "$difficulty" ; then
+	sed -i 's/difficulty=.*/difficulty=$difficulty/' server.properties
+fi
+if test -n "$whitelist" ; then
+	sed -i 's/enforce-whitelist=.*/enforce-whitelist=$whitelist/' server.properties
+	sed -i 's/white-list=.*/white-list=$whitelist/' server.properties
+fi
+if test -e ./libraries; then
+	rm -r ./libraries
+fi
+cp -r ../libraries .
 java -Xmx2G -jar ../server.jar nogui &
 # from https://mbien.dev/blog/entry/stopping-containers-correctly
 # Pass SIGTERM to the jvm to allow graceful shutdown
